+++ Modules/Microsoft365DSC/Modules/M365DSCReport.psm1
@@ -1831,6 +1831,10 @@ function Get-M365DSCResourceKey
     elseif ($mandatoryParameters.Count -gt 0)
     {
         # return all mandatory parameters
+        if ($Resource.ResourceName -eq 'EXOTenantAllowBlockListItems')
+        {
+            $mandatoryParameters = $mandatoryParameters | Where-Object Name -NE 'Action' # Action is not a key property but still mandatory
+        }
         return @($mandatoryParameters.Name)
     }
     elseif ($mandatoryParameters.Count -eq 0)