+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 # UNRELEASED
 
 * AADAuthenticationMethodPolicyFido2
-  * Add Passkey profile handling
+  * Add PassKeyProfile handling
 * AzureRoleAssignmentScheduleRequest
   * Initial Release. [#6945](https://github.com/microsoft/Microsoft365DSC/issues/6945)
 * AzureRoleEligibilityScheduleRequest