+++ CHANGELOG.md
@@ -16,13 +16,19 @@
 * IntuneDeviceConfigurationHealthMonitoringConfigurationPolicyWindows10
   * Fixed an issue with value handling when creating or updating policies.
     FIXES [#6955](https://github.com/microsoft/Microsoft365DSC/issues/6955)
+* SCLabelPolicy
+  * Fixed an issue where setting `AdvancedSettings` failed.
+    FIXES [#6973](https://github.com/microsoft/Microsoft365DSC/issues/6973)
+* SCSensitivityLabel
+  * Fixed an issue where setting `AdvancedSettings` failed.
+    FIXES [#6973](https://github.com/microsoft/Microsoft365DSC/issues/6973)
 * TeamsEmergencyCallingPolicy
   * Added explicit cast to string for `ExternalLocationLookupMode`.
 * M365DSCUtil
   * Removed the internal `Sync-M365DSCParameter` function.
 * MISC
   * Fixed an issue where hardcoded Azure urls were used in multiple resources.
-    FIXES [#6957](https://github.com/microsoft/Microsoft365DSC/issues/6957
+    FIXES [#6957](https://github.com/microsoft/Microsoft365DSC/issues/6957)
 * M365DSCDocGenerator
   * Fixed an issue where no distinction between read and update
     was done for EXO resources.