+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 # UNRELEASED
 
+* EXOTenantAllowBlockListItems
+  * [BREAKING CHANGE] Removed the Microsoft reserved `SubmissionID` parameter.
+  * [BREAKING CHANGE] Updated the property `Action` to be Required instead of Key.
+  * Fixed an issue where comparing resources would compare the wrong instances.
+    FIXES [#6981](https://github.com/microsoft/Microsoft365DSC/issues/6981)
+
+# UNRELEASED
+
 * AADAuthenticationMethodPolicyFido2
   * Add PassKeyProfile handling
 * AzureRoleAssignmentScheduleRequest